Continued integrating the new architecture.  Adjusted the other HMMER programs besides hmmsearch.  This version still does not completely build.
I am currently stalled on fixing hmmalign on modelconfig.c.  In general, 
I still have to figure out how to incorporate the new model configurations
into the new architecture, or vice versa.

I edited Makefile.in so that it is easier to compile a specific customization
of HMMER.  This is kinf of hack, but it seems to work.  We might want to figure
out something better before we release.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/branches/buhler@1412 db3e70e4-12e1-0310-b23f-81a6e3d09ba7