Revisions to create_profmark [J3/30]; monores shuffle is now default, after dires shuffle again fails to adequately shuffle even large sequences; source of negative segments is now tracked; --single option can create single-domain positive set.
In hmmsearch, inserts are now being forced to zero after profile is
read. This is an unclean hack, and we're liable to forget we're doing
this; it should be somewhere more permanent and general, if we're
going to do this. FIXME.

Bugfix in sum score override: was doing the omega calculation twice.

in null2.c, the two functions are now named p7_null2_ByExpectation()
and p7_null2_BySampling(); and the domaindef structure now tracks an
array <n2sc> that gives the null2 correction per target sequence
residue.

<n2sc>, by the way, neatly solves previous issues with calculating the
null2 per-sequence correction for overlapped domains. (We still have
an issue with the sum score, though).



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@2408 db3e70e4-12e1-0310-b23f-81a6e3d09ba7