Conceptual flaw in rescaling forward/backward discovered; rescaling is not guaranteed to work, and under certain circumstances can cause problems.
(That is, underflow is not a problem if the odds are truly supposed to
be zero; but underflow in a state is a problem if that state is
actually used in a high-probability path, and overflow is always a problem.)

backwards_engine now switches on the fly to using its own rescaling
variables if impending overflow is detected. However, this does not
fully solve the problem (and it's actually unclear if it helps
significantly at all!)

the real fix is in the posterior decoding algorithms. A numeric
problem shows up here as an overflow in a product of rescaling ratios;
this overflow is detected, and the routines return <eslERANGE>. I
believe this only happens on unilocal models, in domain
postprocessing, in rare examples where domain identification has
already failed (an envelope still contains >= 2 independent domains),
and I believe these examples will always be repetitive or biased
composition sequence. Therefore, the caller considers envelopes that
generate an <eslERANGE> return code to be ignorable; only a warning is
printed (so we'll know the extent of the problem; we'll see the
warnings, or people will report them to us).

[xref J3/119-121]


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@2575 db3e70e4-12e1-0310-b23f-81a6e3d09ba7