xref J4/47-48. Added documentation/alphaguide: user's guide for the alpha release, in progress. tightened hmmsearch, phmmer (tophits) output. added --textw, --notextw option to hmmsearch; consistent w/ phmmer
IMPORTANT: changed --biasfilter to --nobias options; make use of
biasfilter the default again. This now re-exposes problems that Rob
Finn detected, where HMMER2 finds homologs that H3 misses, because H3
bias filter is overaggressive. The test for this is that --max (and
in particular --nobias) results in the homolog appearing again.)




git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@2695 db3e70e4-12e1-0310-b23f-81a6e3d09ba7