Changed the hmmer programs to build both the serial and threaded processing loops if pthreads is available.  The --cpu flag will control the number of worker threads used by the programs.  If --cpu is zero, the serial version will be run.
The mpi build will also build the serial and threaded processing loops
if pthreads is available.  If the --mpi flag is used the mpi version
will be run.  If not, the --cpu flag will control which version of the
programs will execute.  A runtime error will be displayed if both the
--cpu and --mpi options are used.

The default is to run the threaded version with as many worker threads
as cores.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3012 db3e70e4-12e1-0310-b23f-81a6e3d09ba7