bug #h65 fixed: hmmpress -f corrupts HMM database instead of overwriting it
In hmmpress -f, p7_hmmfile_Open() sees the press'ed database and opens
it, because it's supposed to use the binary db if available; but
unfortunately the db is about to be overwritten by hmmpress -f. Duh.
 
Fix is to implement p7_hmmfile_OpenNoDB(), which opens only the
<hmmfile>, even if a press'ed db is available for it.

Added hmmpress.itest.pl, an integrated test driver that would fail on
this bug. Added it to the testsuite.sqc.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3042 db3e70e4-12e1-0310-b23f-81a6e3d09ba7