Bug #h68 fixed: hmmalign fails on some seqs: Decoding ERANGE error
hmmalign silently fails on some high-scoring sequences because
p7_Decoding() returns an ERANGE error.

The ERANGE error is documented in [J3/119-121] and in the
impl_sse/decoding.c::p7_Decoding() code. Models in unilocal mode may
have numeric overflow on targets containing more than one
"domain". Search progams avoid this problem because of careful domain
postprocessing; it is rare for p7_Decoding() to see a multidomain
target, except in unusual situations that I believe always involve
garbage repetitive sequences, and therefore the failure can be
silently ignored.

But in hmmalign, we are in unilocal mode to begin with, and a user may
provide us with a "multidomain" target.

This needs more thought, but for now, the fix is to detect the
overflow in hmmalign and fail over to slow generic routines.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3048 db3e70e4-12e1-0310-b23f-81a6e3d09ba7