old itest_brute now gets built again by "make check" or "make dev". Added ${ITESTS} targets in Makefiles: compiled integrated tests (freestanding programs solely for testing purposes).
itest_brute does not pass, and is not yet included in automated
testsuite. Difference may be legacy differences between 2007 version
of glocal (when itest_brute was written) and current, after the work
on local was finalized for the PLoS paper. Need to revisit old notes
on the manually enumerated paths in the test.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3068 db3e70e4-12e1-0310-b23f-81a6e3d09ba7