Fixed problems in p7_GTrace() reported by Elena because hmmsim --ls -a fails.
Was assuming Mk->E only; no D_M->E transition was being traced, but
that can happen in a glocal model (though not in a local model). Added
a D_M->E trace step in GTrace() when <gm> isn't a local model.

Fixed a problem in throwing a failed traceback exception from E to a
M_k; test for k==0 after loop ends, not k<0.

Added a trace_Reuse() call in hmmsim -a code (by itself, that was
throwing an exception, regardless of the GTrace() bug).



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3081 db3e70e4-12e1-0310-b23f-81a6e3d09ba7