bug #h71 fixed: hmm->compo calculate wrong, bug in iocc[] calculation.
Elena reports a bug in p7_hmm_CalculateOccupancy. Mean of the insert
geometric is 1/t_IM, not 1/t_II.

This affects the calculation of hmm->compo, the model mean
composition, which is used to construct the two-state HMM used by the
bias filter. Therefore the bugfix had the potential to alter
results. Verified that a benchmark (profmark) is essentially unchanged
-- there is a tiny change, insignificant but unfortunately detrimental
(by about 1%). 

This bugfix may change results, because targers can appear/disappear
due to the alteration in the bias filter.

Added a unit test, utest_composition(), to verify that compo is
correctly calculated -- validation against composition of large set
of sampled emitted seqs.

Also verified that the removal of the OR test from p7_pipeline on 29
June had a similar small detrimental effect on benchmarks. This means
we're backsliding slowly in benchmark performance, and this will need
to be looked at afresh.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3091 db3e70e4-12e1-0310-b23f-81a6e3d09ba7