Bug #h79 fixed: Memory corruption in resized generic P7_GMX  (reported by Elena)
p7_gmx_GrowTo() could get into a situation where it is required to
shrink allocW to be able to set enough valid rows (validR) while
reusing the existing ncells allocation -- but it did not shrink
allocW, resulting in a "successful" return with validR < L+1, leading
to memory corruption.

The equivalent routines in P7_OMX production code are unaffected by
the bug; this only affects the use of P7_GMX, which we only use in
testing.

Unit test for p7_gmx_GrowTo() updated to catch this bug type.
p7_gmx_GrowTo() revised to always resize allocW to M+1 when
row pointers are reset.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/branches/3.0@3268 db3e70e4-12e1-0310-b23f-81a6e3d09ba7