On a suggestion from Bryan Lund to expand the use of the hmmalign --allcol option; instead, make that behavior the default for all output alignments. All output alignments will include all consensus columns, even if the column is all-deletions (gaps).
The original behavior for output MSAs can be achieved using
esl-reformat --mingap.

The hmmalign --allcol option is removed.

The i12-delete-corruption.pl test is now obsolete, because it tests a
behavior (bug #h77) that's only created by MSAs that delete all-gap
consensus columns. Left it in anyway - might need it again someday.

For once, a change that simplifies and shortens the code...



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3277 db3e70e4-12e1-0310-b23f-81a6e3d09ba7