Big commit for all the changes required to make nhmmer work.  nhmmer and all the existing hmmer tools work in my tests, and make check passes.  Making documentation in easel doesn't work ... but that's not the fault of this submission - (a) it wasn't working with a fresh extract from SVN prior to doing this commit.
There are still a few things on tap:
(1) Threads are implemented, but mpi is not. They will be soon.
(2) An implementation exists for impl_sse, but not impl_vmx or impl_dummy.  Also coming soon.
(3) I've run numerous tests of my own, but haven't implemented any tests to be run in "make check".  Clearly I'll need to concoct/implement those.  I'm guessing I'll have a few fine examples once Tom starts beating the tar out of nhmmer.


This commit also fixes the bug in which worker threads on OS X lost the flush-to-zero flag required to avoid denormal slow-down.  The fix is included in hmmscan, hmmsearch, jackhmmer, nhmmer, and phmmer

git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3311 db3e70e4-12e1-0310-b23f-81a6e3d09ba7