Added the first round of support for running a hmmer program as a daemon.
The syntax for the hmmer daemon is:
   hmmd [-p port] <hmmpmg> [options...] database

The only hmmer program currently supported is phmmer. For example:
   ./src/hmmd ./src/phmmer ../../sequences/sprot57

The default port for the daemon to listen on is 41139.

A client program that sends sequences to the daemon and prints
the output is hmmc.  The usage is: hmmc [-i addr] [-p port]
If no address for the server is specified, the loop-back is
used, ie 127.0.0.1.  If no port is specified 41139 is used.

The sequences are read from stdin.  The sequences are in FASTA
foramt and terminated with a // line.  For example:

    >P00999 TIRED_RAT
    GYGCANFDTIAAN
    //




git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/branches/farrar-web@3344 db3e70e4-12e1-0310-b23f-81a6e3d09ba7