Reworked the binary data format shipped back to the client.
The order of the binary data is:
  search status        (HMMD_SEARCH_STATUS)
  search statistics    (HMMD_SEARCH_STATS)
  list of sorted hits  (P7_HIT)
  display information per hit
    list of domains    (P7_DOMAIN)
    alignment          (P7_ALIDISPLAY and display data)

The err_len field in HMMD_SEARCH_STATUS has been replaced
with msg_size.  The msg_size field is the total size of
the binary data following the status, ie stats, hits,
alignments, etc.

Two fields, nres and nnodes have been removed from
HMMD_SEARCH_STATS.

Hits no longer are followed by the hit name, accession
and description.  The name is integer index for that
sequence.

The hit->dcl pointer is an offset to the P7_DOMAIN data
for that hit.  The offset is calculated from the start
of the statistics structure.  The hits domains are
stored contigiously.

The pointers in the P7_ALIDISPLAY structure are offsets
from the start of the first domain structure for that hit.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/branches/farrar-web@3411 db3e70e4-12e1-0310-b23f-81a6e3d09ba7