Added the ability of the master process to recover from a compute node failure during a search.
If more that one node fails, the master node won't try the search again.
Also, if during the retry another compute node fails the search will not
be tried again.  I am assuming that if more than one node goes down or
the search fails twice, there is something wrong with the search.

For the retry, the block the sequences that did not get run will be
distributed amoung the remaining compute nodes and rerun.  The results
from the first and second run will be merged and returned to the client.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/branches/farrar-web@3433 db3e70e4-12e1-0310-b23f-81a6e3d09ba7