bug #h85 fix: hmmsearch and hmmscan give different results for the same model/sequence pair.
The problem was in the configuration of the "bias filter HMM".  A
p7_pli_NewModel() call resets the expected length of the bias filter
HMM's 0 state to a default of 400. A p7_bg_SetLength() call sets the
expected length of both the null1 model and the bias filter HMM's 0
state to the sequence length. In hmmscan, p7_pli_NewModel() is called
as each model is read in the inner target-reading loop, but
p7_bg_SetLength() was being called just once in the outer
query-reading loop. (In three places: serial, thread, and MPI
implementation.) This meant that in hmmscan, the expected length
setting in the biased filter HMM was always at its default of 400. The
more a sequence length deviated from that length, the more likely that
hmmscan would filter it differently than intended. (Thus not
surprising that the report came on a sequence of length 4602.) 



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3440 db3e70e4-12e1-0310-b23f-81a6e3d09ba7