Merged the farrar_web branch back to development trunk. Work on the farrar_web branch will now halt. For now, farrar_web's hmmpgmd is considered stable and ready for production. After some more testing, dev trunk's hmmpgmd will be used instead.
This merges a total of 40 revisions along the branch: -r3343:3435
Conflicts were resolved in Makefile.in (nhmmer and hmmpgmd are both in
PROGS); src/impl_sse/io.c (Michael was hacking in a "3/d" format, see
below); and src/p7_hmm.c (use of threadsafe ctime_r, trunk had
independently implemented this in a better way).

Two other issues arose, summarized as:
 1. watch for buffering problems in hmmpgmd's reading of query sequences;
 2. the current HMM file format is now "3/d".

In more detail:

1. Michael had introduced a bug into esl_sqio_ascii.c::loadbuf(), in
how Easel reads through a stdin pipe. The i17-stdin.pl test fails
(this test is in the trunk, but not the branch). His note indicates he
was trying to fix a problem in how the hmmpgmd daemon reads input from
its client. Apparently fread() buffering prevented the daemon from
receiving the end of a query sequence, because it was still in a
partial buffer.  He introduced a read() call which apparently has a
(side?) effect of flushing the desired buffer. But you can't mix
read() and fread() calls against the same input stream, because they
each keep state information on the stream. I reverted the Easel code
to the original fread()-only version, fixing the bug that the
testsuite detects. However, this may reexpose whatever problem in
hmmpgmd Michael was trying to fix! I don't see it myself; hmmpgmd
seems to work fine, at least using the hmmc2 test client.

2. Michael was hacking in a "3/d" format, only in the SSE impl. I
think he was in the middle of introducing the Knudsen SSVFilter(), and
3/d format meant binary files with SSV data vectors. This created
version mismatch issues, so I upgraded everything to have 3/d as the
current file format.




git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3453 db3e70e4-12e1-0310-b23f-81a6e3d09ba7