Fixed a bug found by Rob in hmmpgmd: it was not respecting bit score thresholds, including --cut_ga.
Problem was that the pipeline tests for pli->hfp != NULL to see if the
pipeline is in SCAN mode, in the block that ends up reading the rest
of the profile and configuring it (incl length distribution params and
bit score thresholds). But in hmmpgmd, all profiles are cached and
hfp==NULL even in SCAN mode.

Changed to test pli->mode == p7_SCAN_MODELS directly, and only test
pli->hfp to see if a call to ReadRest() is needed (if profile is on
disk rather than the hmmpgmd cache). 

Added i19-hmmpgmd-ga.pl test, which looks for this bug. This is the
first hmmpgmd test in the testsuite. We can use it as a model for
future ones. First need something to generate hmmpgmd daemon format
for test seq files though.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3460 db3e70e4-12e1-0310-b23f-81a6e3d09ba7