When very high E-value cutoff (e.g. -E 1000) is specified, nhmmer was correctly reporting hits in the summary section, but giving the following message for each hit in the details section: "[No individual domains that satisfy reporting thresholds (although complete target did)]"
That message was caused by the fact that the domE threshold defaults to 10, is not accepted as an nhmmer flag, yet was (inappropriately) being used to filter "domains" in DNA hits.  This shouldn't be used, since the "domain" part of hmmer's pipeline is only used as a convenience by nhmmer.  This submission causes domain reported/included flags to be set to agree with sequence flags when nhmmer is used. (note: this problem also arose for a very low T score cutoff).

---

This raises a question (TODO):  what happens if hmmsearch is used on DNA sequences?  It currently will present the standard hmmsearch output, including domain reporting.  That disagrees with nhmmer reporting, and seems like we may want to address it.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3476 db3e70e4-12e1-0310-b23f-81a6e3d09ba7