nightly builds are now working in HMMER
ppc-aix-xlc config added to autobuild: power pc, IBM pwr6 platform,
IBM xlc compiler w/ altivec/VMX. This won't build - there are problems
now in the impl_vmx - but it won't hurt to have nightly builds pester
me about that.

TODO: Note that autoconf is not properly configuring CFLAGS for
altivec in the ppc-aix-xlc config; the config is doing it manually.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3477 db3e70e4-12e1-0310-b23f-81a6e3d09ba7