Fixed Rob's segfault in hmmpgmd. Two trivial bugs combined to make one more mysterious one.
first, p7_hmmfile_OpenBuffer() was out of date and only recognizing 3/a and
3/b formats; 3/c and 3/d would be unrecognized (Rob's was current 3/d;
my small HMM tests in testsuite, tutorial are old 3/b).

second, hmmdmstr.c mishandled error return from OpenBuffer(); hfp is
NULL, but hmmdmstr.c tried to report hfp->errbuf, and thus segfaulted.

TODO: OpenBuffer() is not long for this world anyway; scheduled for
replacement w/ new ESL_BUFFER based parsing, unifying all input
sources into one fast interface. (Which will also fix Eric's issues in
Infernal w/ reading H3 format as part of a CM file.)

NOTE: a very handy tool on OS/X: pbcopy. To cut and paste a large
file, where it would be tedious to do the cut with the mouse (because
the text is so large you have to do it in multiple pieces):
  cat file | pbcopy
This puts the contents of "file" in the mac clipboard.  Used this to
paste Rob's HMM example into hmmc2.




git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3483 db3e70e4-12e1-0310-b23f-81a6e3d09ba7