serious thread race condition heisenbug fixed in hmmpgmd.
Farrar's trick to share a mutex across all threads went awry. He
accidentally gave each thread its own mutex, which of course defeats
the purpose of the mutex. The result is that the code for distributing
work was unprotected, causing a race condition. The problem was
essentially invisible to code inspection, because the relevant block
of code looked like it was wrapped in appropriate mutex calls.

The main effect of this is that the same target could get sent to two
threads and worked on in duplicate. This could produce duplicate hits
on output hit lists. (I'd noticed this, and I think Rob and Jody had
also, but it was rare and irreproducible.)

The more serious problems arise because there is a single set of
OPROFILE's, distributed to worker threads; and the length model and
algorithm mode in an OPROFILE are changed during domain
postprocessing. So it is essential that the same OPROFILE is NEVER
NEVER being handled by two threads at the same time.

So in this bug, if one thread went into multidomain region resolution
by stochastic traceback, a more serious flaw could arise.  Stochastic
traceback depends on putting the model temporarily back in multihit
mode, whereas normal domain postprocessing on single-domain regions is
done in unihit mode. If one thread put the model into unihit mode,
clobbering the multihit config in use by stochastic traceback, various
memory-corrupting things could happen, including getting a traceback
stuck in a J state forever (because the E->J probability is 0 in
unihit mode).  The phenotype of the bug is variable, depending on
exact order of execution and when the E-J probability gets clobbered:
segfaults in StochasticTrace are the biggest giveaway, but it could
also produce abort traps, corruption in the malloc heap, and silent
worker thread failures that led to infinite hangs in pthread_join() -
which is the flaw that Rob originally reported.

Because the serious version required not only a race condition, but
additionally the rare case of a model/sequence getting handed off to
stochastic traceback, AND having that stochastic traceback coincide in
time with the other race-condition-deluded thread working on the same
model in unihit mode, AND on exact details of order of execution --
the problem proved impossible to isolate in a reproducible test case.

The fix is simple: to allocate a single mutex, and share the *pointer*
to that mutex with all the threads.

mutter mutter mutter.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3485 db3e70e4-12e1-0310-b23f-81a6e3d09ba7