New default mixture Dirichlet prior for nucleotide models.
This is a 4-component mixture based on the one identified by Eric as outperforming the previous 5-component default. The only difference between this mixture and Eric's is that I've reduced the numerical precision from 6 digits to two, with the intention of avoiding extreme overfitting.

Comparing the two mixtures on Rmark3, the new mixture improves the MER score by 18:

==> r3-nhmmer.default.em1.mer <==
  *family-sum*      294   780   284    10         -
  *summary*         329   780   276    53  1.01584e+06  53.00s

==> r3-nhmmer.c4.t4.rounded.em1.mer <==
  *family-sum*      283   780   274     9         -
  *summary*         311   780   263    48  1.01584e+06  69.00s
  
  
  Xref: ~wheelert/notebook/2011/0325_nhmmer_new_parameters/00README

git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3516 db3e70e4-12e1-0310-b23f-81a6e3d09ba7