Patched over a problem in hmmdwrkr.c reported by Travis. Farrar has a hack in send_results() where he passes <offset> by storing it in the <dcl> pointer of a hit structure. This requires casting an integer value to a pointer. Weird and bad, but functional so long as sizeof(pointer) >= sizeof(offset integer) on both the sending and receiving end. Problem is, he has <offset> as uint32_t at receiving end, but uint64_t in send_results(). That's unsafe on any 32-bit platform, and may therefore generate compiler warnings. Travis saw a warning on his Mac OS/X laptop; I see a warning on Linux in icc ia32 mode.
My patch is simply to make it a uint32_t in send_results(). I also
added a note that this is still an unnecessarily obscure hack, and it
should be revisited. Communicating results in a more efficiently
compressed binary form is already on the books as a to-do, so we'll
revisit this issue then.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3519 db3e70e4-12e1-0310-b23f-81a6e3d09ba7