If 'make check' is interrupted (ctrl-C) while the i19-hmmpgmd-ga.pl test is running, the test would leave zombie processes squatting on the test ports.  This fix captures SIGINT, and tears down the daemons before accepting its doomed fate.  This is achieved via the combination of catch_sigint() and create_kill_script().
I identified this issue because test i19 happens to follow test i18, which is the nhmmer-generic test : I ran 'make check', saw nhmmer failed, and performed a ctrl-C while i19 was running. 

I'm not sure if there are other tests that risk leaving zombies behind, but I'm certain that many of the tests leave temporary files behind when interrupted with SIGINT.  This approach could prove useful to avoid that situation universally.

git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3520 db3e70e4-12e1-0310-b23f-81a6e3d09ba7