New tabular output to support Dfam/Pfam build scripts.
The current Pfam build script calls hmmsearch with both --tblout and --domtblout, then reorganizes the contents of those files for internal consumption. By adding an Xfam-specific output format to hmmsearch and nhmmer, we can simplify the Pfam and Dfam build process.

* Added new function p7_tophits_TabularXfam() to p7_tophits.c. 
* Added commandline argument handling to hmmsearch and nhmmer to support calling that function (--pfamtblout for hmmsearch, --dfamtblout for nhmmer;  moderately different output for the two)

Will eventually push these format changes to jackhmmer and (?) phmmer, but only after the changes are locked in place.

git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3533 db3e70e4-12e1-0310-b23f-81a6e3d09ba7