Location for Easel is now flexible, allowing HMMER to be used within another package that includes its own copy of Easel (i.e. Infernal).
This was achieved by adding "HMMER_ESLDIR" variable to configure.ac,
defining the location of the Easel subdirectory to use, in HMMER's
configure.ac this is defined as "easel", but in Infernal it will be
defined as "../easel" allowing Infernal's HMMER subdirectory to use
the same "easel" directory that Infernal uses.

Makefile.in changes consist of defining "ESLDIR", set as
@HMMER_ESLDIR@, and replacing all cases of explicit defining of the
Easel directory as "easel" with the variable "${ESLDIR}".


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3540 db3e70e4-12e1-0310-b23f-81a6e3d09ba7