Eric pointed out a concern about how reverse-complementing interacts with ReadWindow: if you revcomp a sequence, then use esl_sqio_ReadWindow with a non-zero value for C (indicating that C characters should be kept from preceding sequence), you end up with the first C chars of the new window being revcomp'ed.
I'd written serial_loop in a way that avoided this problem, by copying the sequence before revcomping, then doing revcomp on that copy ... but looking closer at the threaded code (thread_loop), realized this was a problem ... but that there was an even bigger issue: 
thread_loop is responsible for grabbing sequence that can be processed by pipeline_thread; it It does this using read_block, which handles all the complexity of the fact that the database might be a bunch of short sequences, or a bunch of long (>1Mb) sequences, or a mix. For the purposes of this note, assume they're all long sequences, so read_block is essentially just wrapping read_window. In this case, thread_loop() should:
- read a window
- pass that window to pipeline_thread using the esl_workqueue_ReaderUpdate() infrastructure
- read the next window, allowing a C-overlap with the previously-read window
- repeat

What it did instead was just ugly 
- read a window, and pass it along
- get back a processed window (block)
- use that processed window as the template for a ReadWindow() call with overlap C.
The problems:
(1) even if only using 1 thread, this will screw up, as the processed window is revcomped inside pipeline_thread
(2) much worse, if using multiple threads, this will in general result in a "window" being processed that is made up of non-contiguous positions in the database, e.g.
    
   should be 
      db    --------------------------------
      read1 --------
      read2       ----------
      read3               ----------
      read4                       ----------
  
   was :
      db    --------------------------------
      read1 --------
      read2         --------   (doesn't even capture overlap at all)
      read3       --        --------  (creates string of non-contiguous sequence)
      read4               --        --------
  
~~~~

This doesn't appear to typically impact final results, as these overlapping chunks are very small.  In my tests, it typically didn't even impact the # of segments passing the MSV filter ... but it definitely could have a real impact (most likely a loss of a true hit).  

~~~~

* nhmmer.c contains the fix to the problem above, all in the thread_loop function
* p7_pipeline.c is just along for the ride, fixing a couple of documentation/output-spacing errors

git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3562 db3e70e4-12e1-0310-b23f-81a6e3d09ba7