Fixed typos in hmmsearch.c referring to (presumably) old options name: '--tblfp', '--domtblfp' and '--pfamtblfp'. The correct, current names are '--tblout', '--domtblout' and '--pfamtblout'.
This would have only had an impact in the unlikely event that the
output files specified by the user with those options were unable to
be opened for writing (the incorrect option names occurred only in
error message printf statements when the user-specified file name is
output). However, in those cases, I think hmmsearch would have exited
with an error statement that would probably be confusing to the user.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3564 db3e70e4-12e1-0310-b23f-81a6e3d09ba7