Impact of this commit:  (1) capture information needed to succesfully remove duplicate hits (two hits coming      from overlapping windows in nhmmer). Previous implementation depended on hit->acc      as a unique identifier, which was clearly not acceptable, since acc is optional.  (2) reduce the number of hits passing the pipeline stage, such that low-scoring      "domains" (aka hits in nhmmer) don't get uselessly placed onto the tophits list,      as nhmmer was wont to do.
* hmmer.h: add seqidx variable to the P7_HIT struct.  The seqidx for a sequence 
corresponds to its order in the input file (the first sequence is id 0, second is 1,
and so forth) and will be used in nhmmer as a unique sequence identifier. This
will directly impact sorting, which will be used as part of the duplicate-removal
process

* nhmmer.c: changes to serial_loop, thread_loop, pipeline_thread are made to support 
this numbering.

* p7_pipeline.c: p7_Pipeline_LongTarget() changed to support numbering

~~

Also a number of changes related to sorting

* p7_tophits.c: 
      - rename p7_tophits_Sort() to p7_tophits_SortBySortkey(), retaining semantics
      - rename hit_sorter() to hit_sorter_by_sortkey(), retaining semantics
                (but adding a strandedness comparison prior to the position comparison)
      - create new functions p7_tophits_SortBySeqidx() and hit_sorter_by_seqidx(). 
        These functions allow hits to be sorted by hit>seqidx (with fall-through 
        sorting by strand and position)

* nhmmer.c: now call p7_tophits_SortBySeqidx() before RemoveDuplicates, as the 
RemoveDuplicates function will, in an upcoming commit, depend on that sorted list to 
perform a much faster scan for duplicates (rather than comparing all pairs, just 
compare neighbors on the sorted list).  Sorting guarantees that two entries in 
P7_TOPHITS that represent the same hit but are found in different windows will be
adjacent to eachother in the sorted hits list.

* hmmer.h:  in the P7_TOPHITS struct, replaced the single is_sorted variable with the pair
                is_sorted_by_sortkey
                and
                is_sorted_by_seqidx
            (the presence of these new variables is reflected in hmmc2.c, hmmdmstr.c, 
             hmmscan.c, hmmsearch.c, jackhmmer.c, nhmmer.c, and phmmer.c)

* note specifically that I believe my change in hmmc2.c, lines 491-2
        th->is_sorted_by_seqidx  = TRUE;
        th->is_sorted_by_sortkey = FALSE;
is correct, but I'm unclear enough on hmmc2 semantics that this should be verified (Sean?).

~~ 

also in p7_pipeline.c: altered the way domains produced by p7_domaindef_ByPosteriorHeuristics()
are handled in p7_Pipeline_LongTarget(). Previously, all resulting domains were added to 
the tophits list, even though many of them have scores so low that they are guaranteed to 
be filtered out during a later p7_pli_TargetReportable() test. This is appropriate in 
protein search, as below-threshold domains may still be reported as providing additional
support to a high-scoring hit ... but in nhmmer, each "domain" stands on its own as 
a hit, so low-scoring "domains" might as well be thrown out early in the process, so
they don't slow down the sorting/duplicate-removal steps.  This is now done, in a 
conservative fashion (only hits that are guaranteed to be filtered later can be filtered 
at the end of p7_Pipeline_LongTarget).      

~~

Also a slight change to the i19-hmmpgmd-ga.pl testsuite to clean up daemons
prior to calling die(), as zombie daemons cause 'make check' to fail after such 
a death even if the cause of death has been resolved.

git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3572 db3e70e4-12e1-0310-b23f-81a6e3d09ba7