(1) output for --tblout and --dfamtblout now give a "-" when no description is available. I'd previously made it print the empty string "", so there would be no confusion with the printed "-" for hits against the revcomp of the input sequence, but Eric pointed out that this could cause problems for parsers that look to capture a token for desc at the end of a line.  I've now shifted the strand column away from the description column, so I can return the "-" empty description value without inducing confusion about hit strand.
(2) replaced the quadratic-time scan through the hit list for duplicates with a linear
one that depends on having a list sorted by unique seqidx (and strand, and position) as
described in the prior commit. This is the result of discussion about duplicate-removal
with Eric (and mirrors the changes he's already commited to the infernal repository).  
He's done tests that suggest that this will substantially reduce the post-processing run 
time for queries with a large number of hits, for example from several minutes to 
several seconds with hundreds of thousands of hits. 

Also improved the way I indicate that a duplicate is removed, following Eric's lead and
setting the hit's flag to p7_IS_DUPLICATE (new to hmmer.h)


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3573 db3e70e4-12e1-0310-b23f-81a6e3d09ba7