The primary contents of this commit are the bits related to a modified method of computing the composition-bias score modifier, specifically used by nhmmer (though it may be worth adding something like it to the protein search realm).
Specifically, after computing the MSV score, and storing it in the variable 'usc',
the HMMER pipeline approach computes a value 'filtersc' that represents the 
extent to which the composition of the target is biased in the same way as the 
composition of the query model. This score (filtersc) is made up of two parts: 
(1) a sum of scores (log odds ratios) for the full length of the target sequence 
(2) a negative score due to the geometric length-distribution (which is 
    what nullsc is)

A problem with this approach is that a query may be much shorter than the target
sequence, such that there is no way the query aligns to the whole target. The 
null2 score handles this fact by only accruing the positive bias score for the 
envelope of the hit (IIRC), but depends on having already established an envelope,
which we don't yet have early in the pipeline. An extreme example of the problem 
can be seen in a cartoon example of something that can happen in nhmmer:

When msv_longtarget runs, it finds windows around high-scoring diagonals, and merges
overlapping windows. Consider query model Q, of length 500, and with a biased
composition. Suppose msv_longtarget identifies two windows:
  (1) a window W of length 1000 containing a true instance X of the model, and
  (2) a window W5 of length 5000 containing 5 exact copies of W
The MSV score for both windows will be the same, just the score of the highest-scoring
diagonal.  But the sum of scores for the length-5000 window will be 5 times that
of the sum of scores for W.  If those scores are positive (quite likely if the query has 
biased composition), then the filtersc for the longer window will be higher, and 
that longer window may be filtered even though the shorter window isn't.

The same sort of thing can arise in protein, for example with two database 
proteins of different length containing differing # of copies of a query domain.

This issue arises as a major problem in the dmark benchmark I'm building, 
specifically with the Tigger10 query model. That model has 1843 match states, 
and is based on an alignment with composition: 29% A, 29% T, 21% C, 21% D, which
happens to be exactly the composition of the eukaryote-trained dmark background
sequence. Roughly 1/2 of the background sequence passes through the MSV filter
due to this strong bias, and many of the windows are merged to length 7,000 or
greater. The filtersc bias modifiers computed from these windows are mostly 
above 100 nats, and in some cases more than 600 nats (!), and essentially 
all windows are wiped out by the filtersc-based bias filter (>50 planted
positives are found by nhmmer with --nobias, but 0 pass default).

One way to overcome this is to try act on the principle that only the sequence 
being aligned (or some reasonably short window around it) should be considered in
accumulating a score for biased composition. Doing this directly requires knowing 
the range of positions involved in the high scoring MSV (or viterbi or fwd). This
requires backtracking to find alignment boundaries, which may eventually be the 
right thing to do ... but is something I'm choosing to push off to another day
(note: that's a "TODO:")


An ad hoc (read: hack) alternative is to just pick some length that we
expect an MSV hit to be, call it B1, then take the biased-composition score
modifier (filtersc_b1) to be a fraction of the modifier filterc desribed above: 
filtersc_b1 = B1 * filtersc / window_length.  That assumes the bias found by 
p7_bg_FilterScore() is evenly distributed across the full window, which is
probably a conservative estimate of the bias in the actual aligned window.
Similar expected length values can be computed for both viterbi (B2) and 
forward (B3).

I've only tested this approach on the Tigger10 model that raised the issue
in the first place. The result, with some manual tuning, is that 41 hits can be 
recovered while keeping run time a good deal faster than the --nobias
option. The default values are currently 
--B1:  min(window_length, 110)
--B2:  min(window_length, 240)
--B3:  min(window_length, 1000)
Those are clearly not the actual MSV, viterbi, and forward lengths (MSV is 
probably typically only about 20 long, Viterbi maybe 40-60, and Fwd 
alignments are around 100-300 in the output) ... but perhaps the discrepancy 
between these values and the actual aligned length just represents the extent 
to which the bias density in the aligned residues exceeds the bias density of 
the overall window.  

I don't view this as a final solution to early-pipeline bias filtering. At the very 
least, I'll want to study this over a broader range of biased query/db pairs
to understand the sensitivty/speed tradeoffs (that's a TODO:),  and it's 
probably a good idea to push further (into alignment backtracking, or
otherwise). But it's a servicable improvement over what we had before, in 
that it:
(a) recovers almost all of the true hits that are recovered under --nobias, 
    while being substantially faster (when serious bias is present) *
(b) doesn't add to run time for query/db combinations that do not show 
    composition bias (same as --nobias)

* using Tigger10 and Dmark pseudo-chromosomes as an example, --nobias 
took 1724 cpu seconds, while the default modification to bias filtering I've 
described above took 264 cpu seconds, and recovers 41 true positives with 
e-value <0.01 (no false positives), leaving behind only 6 that are found with 
--nobias (best score: 2e-5).   Those few misses are pretty short relative to the
ones passing the pipeline, and truly alignment-length-based score modifier
would almost certianly improve our ability to capture them without unduly 
impacting run time.

--

Other changes to the code include 
* altering the timing of counting when an MSV hit is acheived, to correct an 
  overcounting error
* removal of two lines of unnecessary code for handling overlapping windows 
  found in msvfilter_longtarget()

git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3582 db3e70e4-12e1-0310-b23f-81a6e3d09ba7